+++ assets/javascripts/kitten/components/cards/team-card.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
-import { TeamCardTitle } from 'kitten/components/cards/team-card/title'
-import { TeamCardImage } from 'kitten/components/cards/team-card/image'
-import { TeamCardIcons } from 'kitten/components/cards/team-card/icons'
+import { TeamCardTitle } from './team-card/title'
+import { TeamCardImage } from './team-card/image'
+import { TeamCardIcons } from './team-card/icons'
 
 export class TeamCard extends Component {
   static Image = TeamCardImage