+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update `TeamCard` import links.
+
 ## [1.1.2] - 2018-12-12
 
 Fix: